done_Entity

+++ server/AppTuyenDung/src/main/java/com/doan/AppTuyenDung/entity/Education.java
@@ -8,17 +8,17 @@ import java.time.LocalDateTime;
 @Entity
 @Table(name = "Education")
 public class Education {
-@Id
-@GeneratedValue(strategy = GenerationType.IDENTITY)
-private Long education_id;
-@OneToOne
-@JoinColumn(name = "user_id", nullable = false)
-private User user;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer education_id;
+    @OneToOne
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
 
-private String name_certificate_education;
-private String major;
-private String name_school;
-private LocalDateTime date_start_education;
-private LocalDateTime date_end_education;
-private String description_other;
+    private String name_certificate_education;
+    private String major;
+    private String name_school;
+    private LocalDateTime date_start_education;
+    private LocalDateTime date_end_education;
+    private String description_other;
 }
